Reset database on user change before successful authorization (#64)
